requirements.txt
telethon>=1.16
toml
asyncio
pyre2
systemd
